+++ db/access.php
@@ -42,4 +42,15 @@ $capabilities = [
             'manager' => CAP_ALLOW
         ]
     ],
+
+    'tool/courserating:reports' => [
+        'riskbitmap' => RISK_PERSONAL,
+        'captype' => 'write',
+        'contextlevel' => CONTEXT_COURSE,
+        'archetypes' => [
+            'manager' => CAP_ALLOW,
+            'teacher' => CAP_ALLOW,
+            'editingteacher' => CAP_ALLOW,
+        ]
+    ],
 ];