+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.3] - 2022111300
+### Added
+- Course report with all individual ratings and reviews
+- For Moodle 4.0 - data source for the report builder for all courses ratings and reviews
+
+### Changed
+- Hide "Course rating" field from the course edit form
+
 ## [1.2] - 2022070800
 ### Changed
 - Prevent JS from loading in 'embedded' page layout (fixes bug when used with filter_embedquestion plugin).